+++ .github/workflows/test.yml
@@ -52,8 +52,6 @@ jobs:
         crate:
           - sgx-heap
           - sallyport
-          - untrusted
-          - syscall
           - shim-sgx
           - shim-sev
         profile:
