+++ Cargo.toml
@@ -31,6 +31,7 @@ backend-sev = ["sev", "backend-kvm", "koine"]
 backend-sgx = ["sgx"]
 
 [dependencies]
+sallyport = { path = "internal/sallyport" }
 sev = { version = "0.1", features = ["openssl"], optional = true }
 sgx = { git = "https://github.com/enarx/sgx", rev = "e805b51", features = ["asm", "crypto"], optional = true }
 koine = { git = "https://github.com/enarx/koine", optional = true }
@@ -55,6 +56,7 @@ openssl = "0.10"
 cc = "1.0"
 walkdir = "2"
 protobuf-codegen-pure = "2.22"
+sallyport = { path = "internal/sallyport" }
 
 [dev-dependencies]
 process_control = "3.0"
