+++ internal/shim-sev/Cargo.toml
@@ -10,10 +10,8 @@ name = "shim-sev"
 test = false
 
 [dependencies]
-sallyport = { path = "../sallyport", default-features = false }
+sallyport = { path = "../sallyport" }
 rcrt1 = { git = "https://github.com/enarx/rcrt1", rev = "2205dc7" }
-untrusted = { path = "../untrusted"}
-syscall = { path = "../syscall"}
 compiler_builtins = { version = "0.1", default-features = false, features = [ "mem" ] }
 x86_64 = { version = "0.13", default-features = false, features = ["instructions", "inline_asm"] }
 goblin = { version = "0.3", default-features = false, features = [ "elf64" ] }
