+++ internal/shim-sgx/Cargo.toml
@@ -10,11 +10,9 @@ name = "shim-sgx"
 test = false
 
 [dependencies]
-sallyport = { path = "../sallyport", default-features = false }
-rcrt1 = { git = "https://github.com/enarx/rcrt1", rev = "2205dc7" }
 sgx-heap = { path = "../sgx-heap" }
-syscall = { path = "../syscall"}
-untrusted = { path = "../untrusted"}
+sallyport = { path = "../sallyport" }
+rcrt1 = { git = "https://github.com/enarx/rcrt1", rev = "2205dc7" }
 compiler_builtins = { version = "0.1", default-features = false, features = [ "mem" ] }
 sgx = { git = "https://github.com/enarx/sgx", rev = "5292e53", features = [ "asm" ] }
 goblin = { version = "0.3", default-features = false, features = [ "elf64" ] }
