--- internal/syscall/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "syscall"
-version = "0.1.0"
-authors = ["Harald Hoyer <harald@redhat.com>"]
-license = "Apache-2.0"
-edition = "2018"
-publish = false
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-sallyport = { path = "../sallyport", default-features = false }
-untrusted = { path = "../untrusted"}
-libc = { version = "0.2", features = [] }
-primordial = "0.1"
