--- internal/untrusted/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "untrusted"
-version = "0.1.0"
-authors = ["Harald Hoyer <harald@redhat.com>"]
-license = "Apache-2.0"
-edition = "2018"
-publish = false
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-primordial = "0.1"
